Grammatica, MibbleValidator, MibbleSharp framework switched to .net standard 2.0 for future cross platform compatibility.
JunoSnmp was removed for now.
Added SnmpLextmWrapper for Lextm snmp library
Created basic draft for MibbleBrowser, for poc.

+++ MibbleSharp/NamedNumber.cs
@@ -154,7 +154,7 @@ namespace MibbleSharp
         {
             get
             {
-                if (this.number > int.MaxValue)
+                if (this.number > (BigInteger)int.MaxValue)
                 {
                     return int.MaxValue;
                 }
